requirements.txt
django
crispy-bootstrap5
django-crispy-forms
djangorestframework